requirements.txt
certifi==2020.12.5
cachetools==4.2.1
cffi==1.14.4
chardet==4.0.0
idna==2.10
Flask==1.1.2
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.11
jupyter-console==6.2.0
jupyter-core==4.7.0
jupyter-http-over-ws==0.0.8
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
Keras-Preprocessing==1.1.2
matplotlib==3.3.4
notebook==6.2.0
numpy==1.19.5
Pillow==8.1.2
python-dateutil==2.8.1
pytz==2021.1
pyxdg==0.25
requests==2.25.1
requests-oauthlib==1.3.0
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.4.1
tensorflow-datasets==4.2.0
tensorflow-estimator==2.4.0
tensorflow-examples @ git+https://github.com/tensorflow/examples.git
tensorflow-metadata==0.29.0
urllib3==1.26.3


# yeni eklenen kutuphaneler

pandas==1.1.5
keras==2.8.0
sklearn==0.0